refactor(authlib): Move single-owner packages into their owning plugin
Four authlib packages had exactly one in-tree importer each. Moving
them under the plugin that owns them makes the authlib/ layout honest
about what's shared vs plugin-local, and models the conventional
"plugin owns its internals" pattern for out-of-tree plugins.

Moves (git mv, history-preserving):

  authlib/validation  →  authlib/plugins/jwtvalidation/validation
  authlib/exchange    →  authlib/plugins/tokenexchange/exchange
  authlib/cache       →  authlib/plugins/tokenexchange/cache
  authlib/spiffe      →  authlib/plugins/tokenexchange/spiffe

Each plugin also converts from a flat file in authlib/plugins/ to its
own directory (own package):

  authlib/plugins/jwtvalidation.go   →  authlib/plugins/jwtvalidation/plugin.go
  authlib/plugins/jwtvalidation_identity.go → .../jwtvalidation/identity.go
  authlib/plugins/tokenexchange.go   →  authlib/plugins/tokenexchange/plugin.go

Plugin packages now import plugins.RegisterPlugin / plugins.StatsSource
from the parent plugins package explicitly. main.go adds side-effect
imports for the two moved plugins so their init() runs — same pattern
an external plugin would use.

Kept as shared building blocks:

  authlib/bypass   — any future inbound gate plugin wants path bypass
                     (healthz, readyz, .well-known). Clear multi-plugin
                     future, even though jwt-validation is the only
                     current importer.
  authlib/routing  — two in-tree users (jwt-validation, token-exchange).
                     External token-broker (PR #391) rolled its own
                     router because authlib/routing.ResolvedRoute has
                     token-exchange-shaped fields — separate cleanup
                     deferred as a followup (make the routing
                     mechanism reusable without the schema).
  authlib/auth     — composition layer both gate plugins wrap via
                     auth.Auth. Lingers in authlib/ for now even though
                     it's effectively plugin-internal; natural
                     refactor candidate once a third gate plugin
                     emerges or when auth's Stats type is replaced by
                     a plugin-generic metrics interface.

Test-file reorganization:

  authlib/plugins/plugins_test.go  split into three:
    plugins_test.go (package plugins_test): YAML loader, Stats, Build
      — cross-plugin integration tests, black-box imports of jwtvalidation
      and tokenexchange via side-effect.
    plugins/jwtvalidation/plugin_test.go (package jwtvalidation):
      all JWT Configure/Ready/OnRequest tests + invokeOnRequest helper +
      mockJWTVerifier + newTestJWTValidation.
    plugins/tokenexchange/plugin_test.go (package tokenexchange):
      all TokenExchange tests + invokeOnRequest helper.

  Tests that accessed unexported plugin fields (p.cfg, p.inner,
  p.audienceDeriver) naturally moved into the plugin's own package
  where those fields are in-package.

Invariant checks:

  $ grep -rn 'authlib/(validation|exchange|cache|spiffe)' authbridge/authlib/pipeline/
  (empty — pipeline/ still free of plugin-specific imports)

  $ ls authbridge/authlib/ | grep -E '^(validation|exchange|cache|spiffe)$'
  (empty — old package directories removed)
